Feat : Add SensorData Relative Classes MQTT를 통해 얻은 데이터를 DB에 저장하기 위해 SensorDataDto와 SensorDataService를 추가함.

+++ KEBSmartFarm/src/main/java/com/keb/kebsmartfarm/entity/SensorData.java
@@ -25,6 +25,6 @@ public class SensorData {
 
     private Double humidity;
     private Double temp;
-    private Double illuminance;
+    private int illuminance;
 
 }
